pyproject.toml
[project]
name = "enterprise-auto-documentation"
version = "0.1.0"
description = "AI-powered system that automatically generates and maintains Confluence documentation from GitHub repositories"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "pygithub>=2.2.0",
    "atlassian-python-api>=3.41.0",
    "notion-client>=2.2.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[tool.setuptools.packages.find]
include = ["common*", "services*"]
exclude = ["tests*", "dashboard*", "migrations*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
